requirements.txt
et_xmlfile==2.0.0
lxml==6.0.1
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
xlsxwriter==3.2.5
